requirements.txt
Django~=3.2.5
requests~=2.26.0
PyJWT~=2.1.0
djangorestframework~=3.12.4